qt_gui, removing relative imports, making lint-worthy
