+++ qt_gui/package.xml
@@ -15,12 +15,16 @@
   <url type="bugtracker">https://github.com/ros-visualization/qt_gui_core/issues</url>
   <author>Dirk Thomas</author>
 
-  <buildtool_depend>catkin</buildtool_depend>
+  <buildtool_depend>ament_cmake</buildtool_depend>
 
   <build_depend>qt5-qmake</build_depend>
-  <build_depend>python-qt5-bindings</build_depend>
+  <build_depend>python3-qt5-bindings</build_depend>
 
-  <run_depend version_gte="0.3.0">python_qt_binding</run_depend>
-  <run_depend>python-rospkg</run_depend>
-  <run_depend>tango-icon-theme</run_depend>
+  <exec_depend>python_qt_binding</exec_depend>
+  <exec_depend>python-rospkg</exec_depend>
+  <exec_depend>tango-icon-theme</exec_depend>
+
+  <export>
+    <build_type>ament_cmake</build_type>
+  </export>
 </package>
