+++ qt_gui/qt_gui/application_context.py
