+++ qt_gui/qt_gui/container_manager.py
@@ -30,9 +30,9 @@
 
 from python_qt_binding.QtCore import qDebug, QObject, Qt
 
-from .dock_widget import DockWidget
-from .plugin_descriptor import PluginDescriptor
-from .reparent_event import ReparentEvent
+from qt_gui.dock_widget import DockWidget
+from qt_gui.plugin_descriptor import PluginDescriptor
+from qt_gui.reparent_event import ReparentEvent
 
 
 class ContainerManager(QObject):
@@ -44,7 +44,9 @@ class ContainerManager(QObject):
         self._root_main_window = root_main_window
         self._container_descriptor = PluginDescriptor('__DockWidgetContainer')
         self._container_descriptor.set_action_attributes(
-            self.tr('Container'), self.tr('Container for other dock widgets'), 'folder-new', 'theme')
+            self.tr('Container'), self.tr('Container for other dock widgets'),
+            'folder-new', 'theme')
+
         self._containers = {}
 
     def get_root_main_window(self):
