+++ qt_gui/qt_gui/dock_widget.py
@@ -32,8 +32,8 @@ from python_qt_binding.QtCore import qDebug, QEvent, QPoint, QRect, Qt
 from python_qt_binding.QtGui import QMouseEvent
 from python_qt_binding.QtWidgets import QApplication, QDockWidget
 
-from .dockable_main_window import DockableMainWindow
-from .reparent_event import ReparentEvent
+from qt_gui.dockable_main_window import DockableMainWindow
+from qt_gui.reparent_event import ReparentEvent
 
 
 class DockWidget(QDockWidget):
