+++ qt_gui/qt_gui/dockable_main_window.py
