+++ qt_gui/qt_gui/icon_loader.py
