+++ qt_gui/qt_gui/minimized_dock_widgets_toolbar.py
