+++ qt_gui/qt_gui/plugin_context.py
