+++ qt_gui/qt_gui/plugin_descriptor.py
