+++ qt_gui/qt_gui/plugin_handler_container.py
@@ -28,9 +28,9 @@
 # ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
-from .dock_widget_container import DockWidgetContainer
-from .plugin_handler import PluginHandler
-from .window_changed_signaler import WindowChangedSignaler
+from qt_gui.dock_widget_container import DockWidgetContainer
+from qt_gui.plugin_handler import PluginHandler
+from qt_gui.window_changed_signaler import WindowChangedSignaler
 
 
 class PluginHandlerContainer(PluginHandler):
