+++ qt_gui/qt_gui/plugin_manager_dbus_interface.py
