+++ qt_gui/qt_gui/plugin_provider.py
