+++ qt_gui/qt_gui/ros_package_helper.py
