+++ qt_gui/qt_gui/settings.py
