+++ qt_gui/qt_gui/window_changed_signaler.py
